fix(ci): add missing version.ts file
- Create `src/environments/version.ts` as it is required by the CI environment but missing from the repo.
- This fixes the "Cannot find module './version'" error during e2e tests in CI.
