+++ package.json
@@ -14,9 +14,8 @@
     "affected:lint": "nx affected:lint",
     "start:dev": "nx run-many --target=build --all --configuration development --watch --parallel 5",
     "start:prod": "nx run-many --target=build --all --configuration production --watch --parallel 5",
-    "version-info": "node ./scripts/git-version.mjs .",
-    "build:dev": "npm run version-info && nx run-many --target=build --all --configuration development --parallel 5 && npx tsc src/background.ts --outDir dist/bookmarks --target es2022 --module es2022 --moduleResolution node",
-    "build:prod": "npm run version-info && nx run-many --target=build --all --configuration production --parallel 5 && npx tsc src/background.ts --outDir dist/bookmarks --target es2022 --module es2022 --moduleResolution node"
+    "build:dev": "nx run-many --target=build --all --configuration development --parallel 5 && npx tsc src/background.ts --outDir dist/bookmarks --target es2022 --module es2022 --moduleResolution node",
+    "build:prod": "nx run-many --target=build --all --configuration production --parallel 5 && npx tsc src/background.ts --outDir dist/bookmarks --target es2022 --module es2022 --moduleResolution node"
   },
   "private": true,
   "dependencies": {
