+++ src/app/components/list-view/list-view.component.ts
@@ -4,7 +4,6 @@ import {CdkContextMenuTrigger} from "@angular/cdk/menu";
 
 import { SelectionService, BookmarksFacadeService, TagsService } from "../../services";
 import { OrderByPipe } from "../../pipes/order-by.pipe";
-import { FaviconPipe } from "../../pipes/favicon.pipe";
 import { BookmarkMenuComponent } from '../menus/bookmark-menu/bookmark-menu.component';
 import { FolderIconComponent } from '../folder-icon/folder-icon.component';
 
@@ -17,7 +16,6 @@ import { FolderIconComponent } from '../folder-icon/folder-icon.component';
   imports: [
     DragulaModule,
     OrderByPipe,
-    FaviconPipe,
     CdkContextMenuTrigger,
     BookmarkMenuComponent,
     FolderIconComponent
@@ -142,6 +140,21 @@ export class ListViewComponent implements OnInit, OnChanges {
     } else {
       this.selectionService.selectDirectory(item);
     }
+  }
+
+  // https://developer.chrome.com/docs/extensions/how-to/ui/favicons
+  public getFavicon(item: chrome.bookmarks.BookmarkTreeNode) {
+    if (item?.url != null) {
+      try {
+        const url = new URL(chrome?.runtime?.getURL("/_favicon/") ?? 'https://www.google.com/s2/favicons');
+        url.searchParams.set("pageUrl", item.url);
+        url.searchParams.set("size", "16");
+        return url.toString();
+      } catch (e) {
+        return '';
+      }
+    }
+    return '';
   }
 
   @HostListener('window:keydown', ['$event'])
