--- src/app/components/settings/about/about.component.css
@@ -1,51 +1,0 @@
-.settings-container {
-  padding: 20px;
-  max-width: 800px;
-}
-
-.section {
-  margin-bottom: 30px;
-  padding: 20px;
-  background: var(--card-bg, #fff);
-  border-radius: 8px;
-  box-shadow: 0 1px 3px rgba(0,0,0,0.1);
-}
-
-h3 {
-  margin-top: 0;
-  margin-bottom: 20px;
-}
-
-.info-grid {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.info-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 10px;
-  border-bottom: 1px solid var(--border-color, #eee);
-}
-
-.info-row:last-child {
-  border-bottom: none;
-}
-
-.label {
-  font-weight: 500;
-  color: var(--secondary-text-color, #666);
-}
-
-.value {
-  color: var(--text-color, #333);
-}
-
-code {
-  background: var(--code-bg, #f4f4f4);
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-family: monospace;
-}
