--- src/app/components/settings/about/about.component.html
@@ -1,21 +1,0 @@
-<div class="settings-container">
-  <h2>About</h2>
-
-  <div class="section">
-    <h3>Application Information</h3>
-    <div class="info-grid">
-      <div class="info-row">
-        <span class="label">Version</span>
-        <span class="value">{{ version }}</span>
-      </div>
-      <div class="info-row">
-        <span class="label">Build Date</span>
-        <span class="value">{{ buildDate }}</span>
-      </div>
-      <div class="info-row">
-        <span class="label">Commit Hash</span>
-        <span class="value"><code>{{ commitHash }}</code></span>
-      </div>
-    </div>
-  </div>
-</div>
