--- src/app/components/settings/about/about.component.ts
@@ -1,16 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { environment } from '../../../../environments/environment';
-
-@Component({
-  selector: 'app-about',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './about.component.html',
-  styleUrls: ['./about.component.css']
-})
-export class AboutComponent {
-  version = environment.version;
-  buildDate = environment.date;
-  commitHash = environment.commitHash;
-}
