+++ src/app/components/settings/settings.component.html
@@ -13,9 +13,6 @@
       <a routerLink="/settings/import-export" routerLinkActive="active" class="nav-item">
         Import & Export
       </a>
-      <a routerLink="/settings/about" routerLinkActive="active" class="nav-item">
-        About
-      </a>
     </nav>
     <div class="sidebar-footer">
       <a routerLink="/" class="back-link">← Back to Bookmarks</a>
