+++ src/app/pipes/index.ts
@@ -1,2 +1,1 @@
 export * from './filter-bookmarks.pipe';
-export * from './favicon.pipe';
