+++ src/environments/environment.prod.ts
@@ -1,6 +1,3 @@
-import { versions } from './version';
-
 export const environment = {
-  production: true,
-  ...versions
+  production: true
 };
