+++ src/environments/environment.ts
@@ -1,6 +1,3 @@
-import { versions } from './version';
-
 export const environment = {
-  production: false,
-  ...versions
+  production: false
 };
