+++ src/environments/version.ts
@@ -1,0 +1,4 @@
+export const versions = {
+  version: '0.0.0',
+  revision: 'dev'
+};
